refactor: modified packages names to singular, implemented same error response structure in books validations exceptions
