+++ src/main/java/com/mercadolibro/entity/AppUserRole.java
@@ -1,4 +1,4 @@
-package com.mercadolibro.entities;
+package com.mercadolibro.entity;
 
 import io.swagger.annotations.ApiModel;
 import io.swagger.annotations.ApiModelProperty;
