+++ src/main/java/com/mercadolibro/entity/Book.java
@@ -1,4 +1,4 @@
-package com.mercadolibro.entities;
+package com.mercadolibro.entity;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Getter;
