+++ src/main/java/com/mercadolibro/exception/BookAlreadyExistsException.java
@@ -1,4 +1,4 @@
-package com.mercadolibro.exceptions;
+package com.mercadolibro.exception;
 
 public class BookAlreadyExistsException extends RuntimeException{
     public BookAlreadyExistsException(String message) {
