+++ src/main/java/com/mercadolibro/exception/BookNotFoundException.java
@@ -1,4 +1,4 @@
-package com.mercadolibro.exceptions;
+package com.mercadolibro.exception;
 
 public class BookNotFoundException extends RuntimeException {
    public BookNotFoundException(String message) {
