+++ src/main/java/com/mercadolibro/exception/NoBooksToShowException.java
@@ -1,4 +1,4 @@
-package com.mercadolibro.exceptions;
+package com.mercadolibro.exception;
 
 public class NoBooksToShowException extends RuntimeException {
 }
