+++ src/main/java/com/mercadolibro/exception/ResourceNotFoundException.java
@@ -1,4 +1,4 @@
-package com.mercadolibro.Exception;
+package com.mercadolibro.exception;
 
 public class ResourceNotFoundException extends Exception {
     public ResourceNotFoundException(String message) {
