+++ src/main/java/com/mercadolibro/repository/AppUserRepository.java
@@ -1,6 +1,6 @@
-package com.mercadolibro.repositories;
+package com.mercadolibro.repository;
 
-import com.mercadolibro.entities.AppUser;
+import com.mercadolibro.entity.AppUser;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface AppUserRepository extends JpaRepository<AppUser, Integer> {
