+++ src/main/java/com/mercadolibro/service/impl/BookServiceImpl.java
@@ -3,10 +3,10 @@ package com.mercadolibro.service.impl;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.mercadolibro.dto.BookReqDTO;
 import com.mercadolibro.dto.BookRespDTO;
-import com.mercadolibro.entities.Book;
-import com.mercadolibro.exceptions.BookAlreadyExistsException;
-import com.mercadolibro.exceptions.NoBooksToShowException;
-import com.mercadolibro.exceptions.BookNotFoundException;
+import com.mercadolibro.entity.Book;
+import com.mercadolibro.exception.BookAlreadyExistsException;
+import com.mercadolibro.exception.NoBooksToShowException;
+import com.mercadolibro.exception.BookNotFoundException;
 import com.mercadolibro.repository.BookRepository;
 import com.mercadolibro.service.BookService;
 import org.springframework.beans.factory.annotation.Autowired;
