+++ src/main/java/com/mercadolibro/service/impl/CategoryServiceImpl.java
@@ -3,8 +3,8 @@ package com.mercadolibro.service.impl;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.mercadolibro.dto.CategoryReqDTO;
 import com.mercadolibro.dto.CategoryRespDTO;
-import com.mercadolibro.entities.Category;
-import com.mercadolibro.exceptions.BookNotFoundException;
+import com.mercadolibro.entity.Category;
+import com.mercadolibro.exception.BookNotFoundException;
 import com.mercadolibro.repository.CategoryRepository;
 import com.mercadolibro.service.CategoryService;
 import org.springframework.beans.factory.annotation.Autowired;
